backend/requirements.txt
# Core FastAPI and web framework dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0

# Database dependencies
sqlalchemy>=2.0.20
alembic>=1.12.0
# psycopg2-binary==2.9.9  # PostgreSQL adapter for production (install separately for Railway)

# Google Sheets API dependencies
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.1.0
google-auth>=2.20.0

# Email dependencies
aiosmtplib>=3.0.0
email-validator>=2.0.0
jinja2>=3.1.0  # For email templating

# Environment and configuration
python-dotenv>=1.0.0
pydantic>=2.4.0
pydantic-settings>=2.0.0

# HTTP client for external API calls
httpx>=0.25.0
requests>=2.31.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.3

# Development dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.10.0
flake8>=6.1.0
isort>=5.12.0
mypy>=1.7.0

# Optional: For better development experience
rich>=13.7.0  # For better console output
typer>=0.9.0  # For CLI commands if needed